requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.3
django-crispy-forms==2.0
django-extensions==3.2.3
django-filter==23.2
django-tables2==2.7.0
django-taggit==4.0.0
djangorestframework==3.14.0
djangoviz==0.1.1
factory-boy==3.3.0
Faker==22.0.0
geographiclib==2.0
graphqlclient==0.2.4
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.8.0
Markdown==3.4.3
oauthlib==3.2.2
pycountry==23.12.11
pycparser==2.21
pydotplus==2.0.2
PyJWT==2.7.0
pyparsing==3.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
whitenoise==6.4.0
zipp==3.15.0
